make the purpose, task, and criteria clearer
In the spirit of transparent assignment design, let's be clear about the
purpose, tasks, and criteria for this assignment. The existing
instructions are very clear about the tasks, but don't say much about
why we're doing this or how they'll be assessed. The latter is pretty
simple -- it's a take-home assignment and will be graded as such -- but
let's make that obvious.

Some links to the New Faculty Orientation session that inspired this:

[Transparent design of syllabi and assignments](https://docs.google.com/presentation/d/13baD5RSJwty3CAS7busrVzMC0Fwpc4MuB0E8RHZGd1Q/present#slide=id.g3726348be0e_2_0)

[Assignment prompt checklist](https://docs.google.com/document/d/1HDayUdo9HGYiOvHreugy0BHwKJSuhG3Hh8lQSTYFNvA/edit?tab=t.0).
It says it's for writing assignments, but (1) writing code *is*, in my
opinion, a form of writing; and (2) there is actually nothing whatsoever
specific to writing assignments in the checklist.